backend/requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-openai
python-dotenv
pydantic
aiosqlite<0.22.0
